Refactor DB: byte enums, serial-based stock, indexes
Convert several int typed enum/status fields to byte (Gender, Type, various Status/Payment/DiscountType) across entity classes to reduce storage and reflect small-range enums. Remove ProductAttribute entity and its DbSet; replace ProductVariants.StockQuantity column with a NotMapped Product.StockQuantity computed from ProductSerials (count of available serials) and add Serials navigation. Update HushStoreDbContext model: new defaults for GUID Ids (NEWSEQUENTIALID()), phone column constrained to varchar(20), added indexes (variants, receipts, order details, carts, etc.), decimal column types and new check constraint CK_Vouchers_Quantity. Add migration DatabaseRefactor (Up/Down + designer) and update model snapshot. Breaking changes: run the new migration and backup data before applying (ProductAttributes table dropped and StockQuantity column removed), and update any code that referenced removed ProductAttribute/StockQuantity.

+++ src/Core/Entities/AuthEntities.cs
@@ -13,7 +13,7 @@ namespace PBL3.Core.Entities
         [Required]
         [MaxLength(100)]
         public string FullName { get; set; } = string.Empty;
-        public int Gender { get; set; } // 0: Male, 1: Female, 2: Other
+        public byte Gender { get; set; } // 0: Male, 1: Female, 2: Other
         public DateTime? DateOfBirth { get; set; }
         [MaxLength(500)]
         public string? AvatarUrl { get; set; }
@@ -26,7 +26,7 @@ namespace PBL3.Core.Entities
 
         // D. QUẢN TRỊ & TRẠNG THÁI
         public bool IsActive { get; set; } = true;
-        public int Type { get; set; } // 0: Admin, 1: Employee, 2: Customer
+        public byte Type { get; set; } // 0: Admin, 1: Employee, 2: Customer
 
         // F. REFRESH TOKEN (Lưu trực tiếp trên User, 1-1)
         public string? RefreshToken { get; set; }
